fix(smu,sdt): add explicit ULong cast to sysconf multiplication -- src/sm/sdt/sdtWAExtentMgr.cpp, src/sm/smu/smuProperty.cpp
Applied explicit ULong type casting to sysconf(_SC_PHYS_PAGES) and
sysconf(_SC_PAGESIZE) multiplication.
This follows the same type-casting approach implemented in: src/id/idl/idl.cpp:172.

Note:
[ fix warnings ]

src/sm/sdt/sdtWAExtentMgr.cpp:52:54: warning: comparison of integers of different signs: 'ULong' (aka 'unsigned long') and 'SLong' (aka 'long') [-Wsign-compare]
   52 |     IDE_TEST_RAISE( smuProperty::getMaxTotalWASize() >
      |     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^
   53 |                     ( idlOS::sysconf(_SC_PHYS_PAGES) * idlOS::sysconf(_SC_PAGESIZE) ),
      |                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   54 |                     ERR_ABORT_INTERNAL_MAXSIZE );
      |                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~

src/sm/sdt/sdtWAExtentMgr.cpp:264:30: warning: comparison of integers of different signs: 'ULong' (aka 'unsigned long') and 'SLong' (aka 'long') [-Wsign-compare]
  264 |     IDE_TEST_RAISE( aNewSize >
      |     ~~~~~~~~~~~~~~~~~~~~~~~~~^
  265 |                     ( idlOS::sysconf(_SC_PHYS_PAGES) * idlOS::sysconf(_SC_PAGESIZE) ),
      |                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  266 |                     ERR_ABORT_INTERNAL_MAXSIZE );
      |                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~

src/sm/smu/smuProperty.cpp:3222:33: warning: comparison of integers of different signs: 'ULong' (aka 'unsigned long') and 'SLong' (aka 'long') [-Wsign-compare]
 3222 |     IDE_TEST_RAISE( sNewMaxSize >
      |     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~^
 3223 |                     ( idlOS::sysconf(_SC_PHYS_PAGES) * idlOS::sysconf(_SC_PAGESIZE) ),
      |                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 3224 |                     ERR_ABORT_INTERNAL_MAXSIZE );
      |                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/sm/sdt/sdtWAExtentMgr.cpp
@@ -50,7 +50,7 @@ IDE_RC sdtWAExtentMgr::initializeStatic()
 
 #if defined(_SC_PHYS_PAGES) && defined(_SC_PAGESIZE)
     IDE_TEST_RAISE( smuProperty::getMaxTotalWASize() >
-                    ( idlOS::sysconf(_SC_PHYS_PAGES) * idlOS::sysconf(_SC_PAGESIZE) ),
+                    ( (ULong)idlOS::sysconf(_SC_PHYS_PAGES) * (ULong)idlOS::sysconf(_SC_PAGESIZE) ),
                     ERR_ABORT_INTERNAL_MAXSIZE );
 #endif
 
@@ -262,7 +262,7 @@ IDE_RC sdtWAExtentMgr::resizeWAExtentPool( ULong aNewSize )
 
 #if defined(_SC_PHYS_PAGES) && defined(_SC_PAGESIZE)
     IDE_TEST_RAISE( aNewSize >
-                    ( idlOS::sysconf(_SC_PHYS_PAGES) * idlOS::sysconf(_SC_PAGESIZE) ),
+                    ( (ULong)idlOS::sysconf(_SC_PHYS_PAGES) * (ULong)idlOS::sysconf(_SC_PAGESIZE) ),
                     ERR_ABORT_INTERNAL_MAXSIZE );
 #endif
